+++ primer-api/Makefile
@@ -1,0 +1,29 @@
+# NOTE:
+#
+# Most commands assume you're running this from the top-level `nix
+# develop` shell.
+
+build:
+	cabal build
+
+configure:
+	cabal configure
+
+check:	test
+
+test:
+	cabal test
+
+docs:
+	cabal haddock
+
+clean:
+	cabal clean
+
+bench:
+
+realclean:
+
+deps:
+
+.PHONY: build bench configure test docs clean realclean deps
